Add files via upload
In this, when you enter a number, and then click on the operator that you want to perform on the number, then the number will disappear, however it will still be present and click on the operator and then click on the second number. Once you press the equal sign, the action will be performed and then the result will be displayed.